TAC-583: Fixed the issue where items with quantity more than 0 but less than 1 would not be submitted for partial refunds

+++ public/channelgrabber/orders/jsx/PartialRefund/Service/partialRefundRequest.jsx
@@ -26,7 +26,7 @@ function formatItemsForRequest(items) {
 
     Object.keys(items).forEach((itemId) => {
         const item = items[itemId];
-        const amount = parseInt(item.selectedAmount);
+        const amount = parseFloat(item.selectedAmount);
         if (amount === 0) {
             return false;
         }
